added bundle like interaction behavior to item bag

+++ src/main/java/com/github/elenterius/biomancy/handler/CustomClientEventHandler.java
@@ -11,7 +11,7 @@ import net.minecraftforge.api.distmarker.OnlyIn;
 public final class CustomClientEventHandler {
 	private CustomClientEventHandler() {}
 
-	public static void onEntityEvent(Entity entity, byte eventId) {
+	public static void onEntityEvent(Entity entity, int eventId) {
 		switch (eventId) {
 			case 0:
 				addParticlesAroundEntity(entity.level, ParticleTypes.HAPPY_VILLAGER, entity);
@@ -24,9 +24,7 @@ public final class CustomClientEventHandler {
 		}
 	}
 
-	public static void onWorldEvent(double x, double y, double z, byte eventId) {
-
-	}
+	public static void onWorldEvent(double x, double y, double z, byte eventId) {}
 
 	public static void addParticlesAroundEntity(World level, IParticleData particle, Entity entity) {
 		for (int i = 0; i < 6; i++) {
